[StatesPathFinder] Avoid loops in state sequence
Currently for most of the problems we look at, loops (an edge going into
the same state) are not useful. However, these loops appear in a lot of
state sequences considered by the algorithm and slow it down
significantly. Until we encounter a problem where loops are necessary to
accomplish the task, we should just ignore them.