analyseoptprob propagate intermediate node constr
Propagate the constraints from intermediate nodes as well, so that we
have a higher chance of filtering out invalid state sequences.
Direction of propagation should be from the final node to the first
node, so that we don't miss out the constraints that don't reach the
final node but are still relevant for the first portion of the
transitions.