Try all state sequences before finding new ones
When goal is defined as set of constraints, every state sequence is a
possible solution. The current implementation grows the list of possible
state sequences faster than they are attempted: for every time we add all
neighbors of the front node to queue, we only try out one sequences.
This could be too space consuming if the graph has large branching factor.