Choose only sequences lead to potential goal state
From the list of goal constraints, we can find a list of states that
contain the maximum number of goal constraints. We choose to compute
only state sequences that end in a state that can potentially be a goal
state. However, future work needs to assess/prove that this heuristic
covers all the possible goal states.