Remove code duplication by merging functions
Previously, there are multiple versions of the same function, to cater
for two cases: when goal is defined as a configuration, and when when
goal is defined as a set of constraints. This duplication is not
desirable, as we need to make changes in multiple places if we want to
make a small change in a function. This commit removes the duplication
by merging the functions together and adding checks within the new
functions to cater for the two cases.

Some modifications are needed for analyseOptimizationProblem when goal
is given as a configuration. So far we have made a few, but there may be
more.