Solve when start state is a potential goal state
Previously, we handled this case when goal is defined as a goal
configuration. However, this case may still occur when goal is defined
as a set of constraints. Moreover, the previous method does not store
the loop transition to the list of solutions, making it difficult to
re-try later on.